% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ef}
\alias{ef}
\title{Emission factors from EMEP/EEA}
\format{
A data table with 6431 rows and 22 variables:
\describe{
  \item{Category}{Buses.}
  \item{Fuel}{Diesel, Compressed Natural Gas (CNG) or Biodiesel.}
  \item{Segment}{Urban Buses Midi <=15 t, Urban Buses Standard 15 - 18 t, Urban Buses Articulated >18 t, Coaches Standard <=18 t,
  Coaches Articulated >18 t,Urban CNG Buses,Urban Biodiesel Buses.}
  \item{Euro.Standard}{Conventional, Euro I,Euro II,Euro III,Euro IV,Euro V,Euro VI,EEV.}
  \item{Technology}{NA,SCR,EGR,DPF+SCR.}
  \item{Pollutant}{CO,NOx,VOC,PM,FC,CH4,NH3,N2O.}
  \item{Mode}{NA,Urban Peak,Urban Off Peak,Rural,Highway.}
  \item{Road.Slope}{-0.06,-0.04,-0.02, 0.00,0.02, 0.04,0.06,NA.}   
  \item{Min.Speed}{Minimum speed for emission factor estimation, in km/h.}
  \item{Max.Speed}{Maximum speed for emission factor estimation, in km/h.}
  \item{Alpha, Beta, Gamma, Delta, Epsilon, Zita, Hta, Thita}{Constant parameters.}
  \item{Reduction.Factor}{in \% units.}
  \item{Bio.Reduction.Factor}{in \% units.}
}
}
\source{
\url{https://www.eea.europa.eu/publications/emep-eea-guidebook-2019/part-b-sectoral-guidance-chapters/1-energy/1-a-combustion/road-transport-appendix-4-emission/view}
}
\usage{
ef
}
\description{
Hot emission factors are speed dependant and are expressed in g/km. They differ by fuel, vehicle class
and engine technology. In previous versions of this chapter a number of functions were provided to 
calculate hot emission factors for the different vehicle categories. All these functions are now 
consolidated into a single equation, given by
}
\details{
EF = (Alpha x V^2 + Beta x V + Gamma + Delta / V) / (Epsilon x V^2 + Zeta x V + Eta) x (1 - RF)

where Alpha, Beta, Gamma, Delta, Epsilon, Zera, Eta are constant parameters that depends on
pollutant, speed, technology, vehicle category.

The emissions factors are derived from the EMEP/EEA air pollutant emission inventory guidebook (formerly 
called the EMEP CORINAIR emission inventory guidebook). The guide provides guidance on estimating emissions 
from both anthropogenic and natural emission sources.
}
\keyword{datasets}
