% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_europe.R
\name{ef_europe}
\alias{ef_europe}
\title{Emission factor dependent on speed by EMEP/EEA}
\usage{
ef_europe(
  speed,
  veh_type,
  euro,
  pollutant,
  fuel = "Diesel",
  tech = "SCR",
  slope = 0,
  load = 0.5,
  fcorr = 1,
  as_list = TRUE
)
}
\arguments{
\item{speed}{units; Speed in 'km/h'.}

\item{veh_type}{character; Bus type, classified in "Urban Buses Midi <=15 t","Urban Buses Standard 15 - 18 t"
"Urban Buses Articulated >18 t",  "Urban Buses Diesel Hybrid", "Urban CNG Buses",  "Urban Biodiesel Buses".}

\item{euro}{character; Euro period of vehicle, classified in "Conventional", "I", "II",
"Euro III", "IV", "V", "VI", and "EEV".}

\item{pollutant}{character; Pollutant, classified in "FC","CO2","CO","NOx","VOC","PM10","EC","CH4","NH3","N2O"}

\item{fuel}{character; Fuel type, classified in "Diesel","Diesel Hybrid ~ Diesel","Diesel Hybrid ~ Electricity",
"CNG","Biodiesel"}

\item{tech}{character; Technology, classified in "SCR", "EGR", and "DPF+SCR" (for Euro VI). Default is "SCR". There are 
no available technology associated with euro standards "Conventional", "I", "II" and "III".}

\item{slope}{numeric; Slope gradient, classified in -0.06, -0.04, -0.02, 0.00, 0.02, 0.04 and 
0.06. Negative gradients means downhills and positive uphills. Default is 0.0.}

\item{load}{numeric; Load ratio, classified in 0.0, 0.5 and 1.0. Default is 0.5.}

\item{fcorr}{numeric; Correction based on fuel composition. The length must be one per
each euro standards. Default is 1.0.}

\item{as_list}{logical; Returns emission factors as a list, instead of data.table format. Default is TRUE.}
}
\value{
emission factors in units 'g/km' (list or a data.table).
}
\description{
Returns a list or data.frame of emission factors for buses based on EMEP/EEA.
Function is based on values from [EMEP/EEA air pollutant emission inventory guidebook 2019](https://www.eea.europa.eu/themes/air/air-pollution-sources-1/emep-eea-air-pollutant-emission-inventory-guidebook).
Estimates expressed in units 'g/km'.
}
\examples{
set.seed(1335)
ef_europe(speed = units::set_units(rnorm(100,50,5),"km/h"),
                veh_type = c("Urban Buses Standard 15 - 18 t","Urban Buses Articulated >18 t"),
                euro = c("IV","V"),
                pollutant = c("CO2","NOx"),
                fuel = "Diesel" ,
                tech =  c("SCR","EGR"),
                slope = 0.0,
                load = 0.5,
                fcorr = 1,
                as_list = TRUE)
}
