% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_usa.R
\name{ef_usa}
\alias{ef_usa}
\title{Running exhaust emissions factors from United States (EMFAC2017 model)}
\source{
\url{https://arb.ca.gov/emfac/}
}
\usage{
ef_usa(
  pollutant,
  calendar_year,
  model_year,
  speed,
  fuel = "Diesel",
  as_list = TRUE
)
}
\arguments{
\item{pollutant}{character; Pollutants: Carbon monoxide (CO), Nitrogen oxides (NOx), 
Hydrocarbons as TOG (total organic gases), ROG (reactive organic gases), 
THC (total hydrocarbon), or CH4 (methane).
Particulate matter as particulate matters 10 microns or less in diameter (PM10), 
and particulate matters 2.5 microns or less in diameter (PM2.5),
Sulfur oxides (SOx), Carbon Dioxide (CO2),
Nitrous Oxide (N2O), and Methane (CH4).}

\item{calendar_year}{numeric; Calendar Year between 2010 - 2020. Year in which the emissions
inventory is estimated, in order to consider the effect of degradation.}

\item{model_year}{numeric; Model year of vehicle.}

\item{speed}{units; Speed in 'km/h'; Emission factor are returned in speed intervals 
such as "5-10", "10-15", "15-20", "20-25", "25-30", "30-35", "35-40", "40-45", "45-50",
 "50-55", "55-60", "60-65", "65-70", "70-75", "75-80", "80-85", "85-90", ">90" mph (miles/h).}

\item{fuel}{character; Type of fuel: 'Diesel','Gasoline','Natural Gas'. Default is 'Diesel'.}

\item{as_list}{logical; Returns emission factors as a list, instead of data.table format. Default is TRUE.}
}
\value{
data.table; Emission factors in units 'g/km' by speed and model_year.
}
\description{
Returns a vector or data.frame of emission factors for buses based on 
values from the [California EMission Factor model (EMFAC2017)](https://arb.ca.gov/emfac/).
Estimates expressed in units 'g/km'.
}
