% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_usa_emfac.R
\name{ef_usa_emfac}
\alias{ef_usa_emfac}
\title{Running exhaust emissions factors from United States (EMFAC2017 model)}
\source{
\url{https://arb.ca.gov/emfac/}
}
\usage{
ef_usa_emfac(
  pollutant,
  calendar_year,
  fuel = "D",
  model_year,
  speed,
  as_list = TRUE
)
}
\arguments{
\item{pollutant}{character; Pollutants: CH4(Methane), CO(Carbon Monoxide), 
CO2(Carbon Dioxide), N2O(Nitrous Oxide), NOx(Oxides of Nitrogen),
 PM10(Primary Exhaust PM10 - Total), PM25(Primary Exhaust PM2.5 - Total), SOX(Oxides of Sulfur),
 TOG(Total Organic Gases), ROG (Reactive Organic Gases)}

\item{calendar_year}{numeric; Calendar Year between 2015 - 2022. Year in which the emissions
inventory is estimated, in order to consider the effect of degradation.}

\item{fuel}{character; Type of fuel: 'D' (Diesel),'G' (Gasoline),'CNG' (Compressed Natural Gas). Default is 'D'.}

\item{model_year}{numeric; Model year of vehicle.}

\item{speed}{units; Speed in 'km/h'; Emission factor are returned in speed intervals 
such as "5-10", "10-15", "15-20", "20-25", "25-30", "30-35", "35-40", "40-45", "45-50",
 "50-55", "55-60", "60-65", "65-70", "70-75", "75-80", "80-85", "85-90", ">90" mph (miles/h).}

\item{as_list}{logical; Returns emission factors as a list, instead of data.table format. Default is TRUE.}
}
\value{
List. Emission factors in units 'g/km' by speed and model_year.
}
\description{
Returns a vector or data.frame of emission factors for buses based on 
values from the [California EMission Factor model (EMFAC2017)](https://arb.ca.gov/emfac/).
Estimates expressed in units 'g/km'.
}
\examples{
 ef_usa_emfac(pollutant = c("CO","PM10"),
        calendar_year = 2019,
        model_year = 2015,
        speed = units::set_units(1:100,"km/h"),
        fuel = "D",
        as_list = TRUE)
        
}
