% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emfac}
\alias{emfac}
\title{EMFAC
Describe}
\format{
A data table with xxx rows and 22 variables:
\describe{
  \item{calendar_year}{calendar_year}
  \item{fuel}{fuel}
  \item{model_year}{model_year}
  \item{speed}{speed}
}
}
\source{
\url{https://www.eea.europa.eu/publications/emep-eea-guidebook-2019/part-b-sectoral-guidance-chapters/1-energy/1-a-combustion/road-transport-appendix-4-emission/view}
}
\usage{
emfac
}
\description{
EMFAC
Describe
}
\keyword{datasets}
