% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emi_do_dt.R
\name{emi_to_dt}
\alias{emi_to_dt}
\title{Convert emissions estimates from list to data.table format}
\usage{
emi_to_dt(emi_list, emi_vars, veh_vars, pol_vars, segment_vars = NULL)
}
\arguments{
\item{emi_list}{List; Emissions list.}

\item{emi_vars}{Character; Name of emissions data on 'emi_list'}

\item{veh_vars}{Character; Attributes of vehicle data on 'emi_list'}

\item{pol_vars}{Character; Attributes of pollutants data on 'emi_list'}

\item{segment_vars}{Character; Attributes of segment link data on 'emi_list'}
}
\value{
data.table.
}
\description{
Read emissions in list format and export to data.table format.
}
\examples{
set.seed(1335)
dist = units::set_units(rnorm(100,0.250,0.03),"km")
ef <- ef_brazil(pollutant = c("CO2","NOx"),
                veh_type = "BUS_URBAN_D", 
                years = 2016)

ef <- ef_europe(speed = units::set_units(rnorm(100,50,5),"km/h"),
                veh_type = c("Urban Buses Standard 15 - 18 t","Urban Buses Articulated >18 t"),
                euro = c("IV","V"),
                pollutant = c("CO2","NOx"),
                fuel = "Diesel" ,
                tech =  c("SCR","EGR"),
                slope = 0.0,
                load = 0.5,
                fcorr = 1,
                as_list = TRUE)

emi <- emis(fleet_composition =  c(0.7,0.3),
            dist = dist,
            ef = ef,
            aggregate = FALSE,
            as_list = TRUE)
              
# to data.table
dt <- emi_to_dt(emi_list = emi,
                emi_vars = "emi", # "emi" var
                veh_vars = c("veh_type","euro","fuel","tech"), # "veh_type"
                pol_vars = "pollutant", # pollutant
                segment_vars = c("slope","load")) # NULL

}
