% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis.R
\name{emis}
\alias{emis}
\title{Estimate hot exhaust emissions}
\usage{
emis(pol_list, input_folder, output_folder = ".", overwrite = TRUE)
}
\arguments{
\item{pol_list}{Character; Vector of list of pollutants}

\item{input_folder}{Character;}

\item{output_folder}{Character;}

\item{overwrite}{Logical;}
}
\description{
Transform intervals of GPS data of constant speed into
a linestring segment, in order to reduce data size of observations.
In Curitiba, for the sake of simplicity, I'm adding:
}
