% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_grid.R
\name{emis_grid}
\alias{emis_grid}
\title{Street emissions into grid}
\usage{
emis_grid(data, emi, grid, time_class = "all periods", time_column)
}
\arguments{
\item{data}{Data.table; Data.table with emissions and departure_time data.}

\item{emi}{Character; Columns names of emissions information.}

\item{grid}{Spatial (sf polygon); Grid cell data to allocate emissions}

\item{time_class}{Character; Emissions can be aggregated by all periods (Default),
'hour' or 'hour-minute'.}

\item{time_column}{Vector; Column name of 'departure_time' information (from data input). 
Only used when 'hour' or 'hour-minute' is selected.}
}
\value{
Spatial data (emissions into grid cells).
}
\description{
Aggregate emissions proportionally into grid cells, by performing an intersection 
operation between emissions data in linestring format and grid cells. It considers aggregation 
per grid cells and
}
