% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_post.R
\name{emis_post}
\alias{emis_post}
\title{Emissions post-processsing}
\usage{
emis_post(data, emi, time_class, time_column = "departure_time")
}
\arguments{
\item{data}{Data.table; Data.table with emissions and departure_time data.}

\item{emi}{Character; Columns names of emissions information.}

\item{time_class}{Character; Emission can be aggregated into 'hour' and'hour-minute'. 
Missing input (Default) no aggregation without considering time consideration.}

\item{time_column}{Vector; Vector containing 'departure_time'. 
Only used when 'hour' or 'hour-minute' is selected.}
}
\value{
Units (in g); emissions
}
\description{
Aggregate emissions generated by emis function by time, 
vehicle type and segment link (spatial)
}
