% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_do_dt.R
\name{emis_to_dt}
\alias{emis_to_dt}
\title{Convert estimates from list to data.table format}
\usage{
emis_to_dt(
  emi_list,
  emi_vars = "emi",
  veh_vars = "veh_type",
  pol_vars = "pollutant",
  segment_vars = NULL
)
}
\arguments{
\item{emi_list}{List; Emissions list.}

\item{emi_vars}{Character; data.frame names of 'emi_list' object attributed to emissions or
emission factors. Default is 'emi'.}

\item{veh_vars}{Character; data.frame names of 'emi_list' object attributed to vehicle
characteristics. Default is 'veh_type'.}

\item{pol_vars}{Character; data.frame names of 'emi_list' object attributed to pollutants.
Default is 'pollutant'.}

\item{segment_vars}{Character; data.frame names of 'emi_list' object attributed to the
road segments. Default is NULL.}
}
\value{
data.table.
}
\description{
Read estimates generated by the \code{\link{emis}} or emission factor
functions (e.g. \code{\link{ef_brazil_cetesb}}), , and convert to a data.table format.
The function allows one to aggregate emissions or emission factors by time of the day,
vehicle type or road segment (spatial).
}
\examples{
if (interactive()) {
library(data.table)
library(magrittr)
gtfs_file <- system.file("extdata/irl_dub/irl_dub_gtfs.zip", package = "gtfs2emis")
gtfs <- gtfstools::read_gtfs(gtfs_file)
# Keep single trip
gtfs <- gtfstools::filter_by_trip_id(gtfs
                                     , trip_id = c('619.3.60-40-d12-1.224.O'
                                                   ,"10812.1.60-13-d12-1.26.I"))
# Transport model
  tp_model <- transport_model(gtfs_data = gtfs,
                              spatial_resolution = 100,
                              parallel = TRUE)
# emission model
fleet_df <- read.csv(system.file("extdata/irl_dub/irl_dub_fleet.txt"
                                 , package = "gtfs2emis"))

emi_list <- emission_model(tp_model = tp_model
                           , ef_model = "ef_europe_emep"
                           , fleet_data = fleet_df
                           , pollutant = c("CO2","PM10"))
# View emission data
names(emi_list)
emis_to_dt(emi_list
           ,veh_vars = c("veh_type","euro","fuel","tech")
)
emis_to_dt(emi_list
           ,veh_vars = c("veh_type","euro","fuel","tech")
           ,segment_vars = c("slope","load")
)
emis_to_dt(emi_list)

emis_to_dt(emi_list
           ,emi_vars = "emi"
           ,veh_vars = c("veh_type","euro","fuel","tech")
           ,pol_vars = "pollutant"
           ,segment_vars = c("slope","load")
)
# View Emission Factor data
emis_to_dt(emi_list
           ,emi_vars = "EF")
emis_to_dt(emi_list
           ,emi_vars = "EF"
           ,veh_vars = c("veh_type","euro","fuel","tech")
           ,segment_vars = c("slope","load"))

# Emission factor model
list_ef_br <- ef_brazil_cetesb(pollutant = c("CO","PM10","CO2")
                               ,veh_type = c("BUS_URBAN_D","BUS_MICRO_D")
                               ,model_year = c(2010,2015)
                               ,as_list = TRUE)
emis_to_dt(list_ef_br
           ,emi_vars = "EF")
emis_to_dt(list_ef_br
           ,emi_vars = "EF"
           ,veh_vars = c("veh_type","model_year"))
}
}
