% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_cor.R
\name{fuel_cor}
\alias{fuel_cor}
\title{Correction factor equation}
\usage{
fuel_cor(pollutant, den, s, pah, cn, t95)
}
\arguments{
\item{pollutant}{character; Pollutant classified in "CO", "VOC", "NOx" or "PM".}

\item{den}{numeric; Density at 15 °C (units kg/m3).}

\item{s}{numeric; Sulphur content in ppm.}

\item{pah}{numeric; Polycyclic aromatics content in \%.}

\item{cn}{numeric; Cetane number.}

\item{t95}{numeric; Back end distillation in degrees C.}
}
\value{
numeric; fuel correction factor.
}
\description{
Relations between emissions and fuel properties for diesel heavy-duty vehicles based on EMEP/EEA. Function based on
values from the [EMEP/EEA air pollutant emission inventory guidebook 2019](https://www.eea.europa.eu/themes/air/air-pollution-sources-1/emep-eea-air-pollutant-emission-inventory-guidebook).
Estimates are dimensionless.
}
