% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps2line.R
\name{gps_as_sflinestring}
\alias{gps_as_sflinestring}
\title{Converts a GPS-like data.table to a MultiLineString Simple Feature (sf) object}
\usage{
gps_as_sflinestring(input_file, crs = 4326)
}
\arguments{
\item{input_file}{A data.table with timestamp data.}

\item{crs}{A Coordinate Reference System. The default value is 4326 (latlong WGS84).}
}
\value{
A simple feature (sf) object with MultiLineString data.
}
\description{
Every interval of GPS data points of constant speed for each trip_id is
 converted into a linestring segment.
}
\examples{
cur_read <- gtfs2gps::read_gtfs(system.file("extdata/gtfs_cur.zip", package ="gtfs2emis"))
cur_gtfs <- gtfs2gps::filter_by_shape_id(gtfs_data = cur_read,shape_ids = c(1708))
cur_gtfs <- gtfs2gps::gtfs2gps(gtfs_data = cur_gtfs)
cur_gpslines <- gps_as_sflinestring(input_file = cur_gtfs) 
plot(cur_gpslines$geometry)

}
