% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_to_linestring.R
\name{gps_to_linestring}
\alias{gps_to_linestring}
\title{Convert GPS to Lines-data type}
\usage{
gps_to_linestring(
  input_filepath,
  output_filepath,
  fleet_data,
  overwrite = TRUE
)
}
\arguments{
\item{input_filepath}{Character; location of exported GPS files}

\item{output_filepath}{Character; location for export Line-data type}

\item{fleet_data}{Data.table; fleet data associated with information of age, engine size,
weight gross total and shape_id's. This input associated with shape_id's allows the function to relate each
shape_id to a vehicle, by creating columns of fleet specification in the output.
Missing entry means no fleet allocation}

\item{overwrite}{Logical; overwrites files in the output_filepath}
}
\description{
Transform intervals of GPS data of constant speed into
a linestring segment, in order to reduce data size of observations.
}
