% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs2emis.R
\name{gtfs2emis}
\alias{gtfs2emis}
\title{Estimate hot exhaust emissions}
\usage{
gtfs2emis(
  gtfs_data,
  fleet_composition,
  ef,
  aggregate = TRUE,
  prefix = NULL,
  as_list = TRUE
)
}
\arguments{
\item{gtfs_data}{A path to a GTFS.zip file, or a GTFS data represented as a 
list of data.tables generated by the `gtfs2gps` package.}

\item{fleet_composition}{vector; Fleet composition, which is a distribution 
of fleet based on frequency. If there is only one.}

\item{ef}{list or data.table;    Emission factors.}

\item{aggregate}{logical; if TRUE (default) emissions are aggregated by pollutant.}

\item{prefix}{character; Add prefix into emissions names. Missing parameter (default)
means empty prefix.}

\item{as_list}{logical; if TRUE (default) emissions are returned inside 'ef' list.}
}
\value{
units ('g'); emissions per link.
}
\description{
Calculate hot exhaust emissions.
}
\examples{
            
}
