% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs2gps.R
\name{gtfs2gps}
\alias{gtfs2gps}
\title{Convert GTFS to GPS-like data given a spatial resolution}
\usage{
gtfs2gps(raw_gtfs, filepath, filter_weekdays = TRUE)
}
\arguments{
\item{raw_gtfs}{Character; a path to a GTFS file to be converted to GPS}

\item{filepath}{Character; output file path.}

\item{filter_weekdays}{Logical; When TRUE (default) is filter the gtfs file, removing the
trips operating only saturday or sunday.}
}
\description{
Convert GTFS data to GPS format by sampling points using a spatial resolution. The
function uses functions from gtfs2gps package, than read gtfs files, 
filter by week day (default option), by a fixed spatial resolution of 15 m.
}
