% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_emfac_db}
\alias{usa_emfac_db}
\title{EMFAC
Running exhaust emissions factors from EMFAC2017 model}
\format{
A data.table with 106920 rows and 7 variables:
\describe{
\item{pol}{Character; Pollutants: CH4(Methane), CO(Carbon Monoxide), 
CO2(Carbon Dioxide), N2O(Nitrous Oxide), NOx(Oxides of Nitrogen),
 PM10(Primary Exhaust PM10 - Total), PM25(Primary Exhaust PM2.5 - Total), SOX(Oxides of Sulfur),
 TOG(Total Organic Gases), ROG (Reactive Organic Gases)}
  \item{calendar_year}{Numeric; Calendar Year between 2010 - 2020}
  \item{fuel}{character; Type of fuel: 'D' (Diesel),'G' (Gasoline),'CNG' (Compressed Natural Gas).}
  \item{model_year}{Model year.}
  \item{speed}{Units; Speed in 'km/h'; Emission factor are returned in speed intervals 
 such as "5-10", "10-15", "15-20", "20-25", "25-30", "30-35", "35-40", "40-45", "45-50"
  "50-55", "55-60", "60-65", "65-70", "70-75", "75-80", "80-85", "85-90", ">90" mph (miles/h)}
}
}
\source{
\url{https://www.eea.europa.eu/publications/emep-eea-guidebook-2019/part-b-sectoral-guidance-chapters/1-energy/1-a-combustion/road-transport-appendix-4-emission/view}
}
\usage{
usa_emfac_db
}
\description{
EMFAC
Running exhaust emissions factors from EMFAC2017 model
}
\keyword{datasets}
