% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_moves_db}
\alias{usa_moves_db}
\title{MOVES
Running exhaust emissions factors from United States (MOVES3 model)}
\format{
A data.table:
\describe{
\item{pollutant}{character; Pollutants: Carbon monoxide (CO), Nitrogen oxides (NOx), 
  Hydrocarbons as TGH (Total Gaseous Hydrocarbons), EC (energy consumption), 
  THC (total hydrocarbon), or CH4 (methane).
  Particulate matter as particulate matters 10 microns or less in diameter (PM10), 
  and particulate matters 2.5 microns or less in diameter (PM2.5),
   Carbon Dioxide (CO2), and Methane (CH4).}
\describe{fuel}{character; Type of fuel: 'D' (Diesel),'G' (Gasoline),'CNG' (Compressed Natural Gas).}
\describe{model_year}{numeric; Model year of vehicle.}
\describe{speed}{units; Speed in 'km/h'; Emission factor are returned in speed intervals 
 such as " - 2.5", "2.5 - 7.5", "7.5 - 12.5", "12.5 - 17.5", "17.5 - 22.5", "22.5 - 27.5",
 "27.5 - 32.5","32.5 - 37.5","37.5 - 42.5","42.5 - 47.5","47.5 - 52.5",
 "52.5 - 57.5", "57.5 - 62.5", "62.5 - 67.5", "67.5 - 72.5", ">72.5" mph (miles/h).}
 }
}
\source{
\url{https://www.epa.gov/moves}
}
\usage{
usa_moves_db
}
\description{
Data.frame of emission factors for buses based on 
values from the [MOVES3 Model](https://www.epa.gov/moves).
Estimates expressed in units 'g/km'.
}
\keyword{datasets}
