% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_euro_scaled.R
\name{ef_euro_scaled}
\alias{ef_euro_scaled}
\title{Scaled emission factors from European (EMEP/EEA) methodology}
\usage{
ef_euro_scaled(
  ef_local,
  speed,
  veh_type,
  euro,
  fuel,
  pollutant,
  SDC = 34.12,
  tech = "SCR",
  slope = 0,
  load = 0.5,
  fcorr = 1
)
}
\arguments{
\item{ef_local}{units; Local emission factors, in units 'g/km'.}

\item{speed}{units; Speed in 'km/h'.}

\item{veh_type}{character; Bus type, classified in "Ubus Midi <=15 t","Ubus Std 15 - 18 t","Ubus Artic >18 t",
"Coaches Std <=18 t","Coaches Artic >18 t".}

\item{euro}{character; Euro period of vehicle, classified in "Conventional", "I", "II", "III", "IV", "V", "VI", and "EEV".}

\item{fuel}{character; Fuel type, classified in "D" (Diesel),"DHD" (Diesel Hybrid ~ Diesel),
"DHE" (Diesel Hybrid ~ Electricity), "CNG" (Compressed Natural Gas), "BD" (Biodiesel).}

\item{pollutant}{character; Pollutant, classified in "FC","CO2","CO","NOx","VOC","PM10","EC","CH4","NH3","N2O". "FC" means Fuel Consumption.}

\item{SDC}{numeric; Speed of the driving cycle in 'km/h'.}

\item{tech}{character; After treatment technology, classified in "SCR" (Selective Catalytic Reduction), 
"EGR" (Exhaust Gas Recirculation), and "DPF+SCR" (Diesel Particulate Filter + SCR, for Euro VI). Default is "SCR" for "IV" and "V".}

\item{slope}{numeric; Slope gradient, classified in -0.06, -0.04, -0.02, 0.00, 0.02, 0.04 and 
0.06. Negative gradients means downhills and positive uphills. Default is 0.0.}

\item{load}{numeric; Load ratio, classified in 0.0, 0.5 and 1.0. Default is 0.5.}

\item{fcorr}{numeric; Correction based on fuel composition. The length must be one per
each euro standards. Default is 1.0.}
}
\value{
list. Emission factors in units 'g/km'.
}
\description{
Estimates scaled emission factor for buses based on EMEP/EEA
}
\examples{
temp_ef_br <- ef_brazil(pollutant = c("CO","PM","CO2","CH4","NOx"),
                        veh_type = "BUS_URBAN_D",
                        years = 2015,
                        as_list = TRUE)

temp_ef_europe <- ef_europe(speed = units::set_units(1:100,"km/h"),
                            veh_type = "Ubus Std 15 - 18 t",
                            euro = "IV",fuel = "D",tech="SCR",
                            pollutant = c("CO","PM10","CO2","CH4","NOx")) 

temp_ef_scaled <- ef_euro_scaled(ef_local = temp_ef_br,
                                 speed = units::set_units(1:100,"km/h"),
                                 veh_type = "Ubus Std 15 - 18 t",
                                 euro = "IV",fuel = "D",tech="SCR",
                                 pollutant = c("CO","PM10","CO2","CH4","NOx"))

ef_europe(speed = units::set_units(rnorm(100,50,5),"km/h"),
                veh_type = c("Urban Buses Standard 15 - 18 t","Urban Buses Articulated >18 t"),
                euro = c("IV","V"),
                pollutant = c("CO2","NOx"),
                fuel = "Diesel" ,
                tech =  c("SCR","EGR"),
                slope = 0.0,
                load = 0.5,
                fcorr = 1,
                as_list = TRUE)
}
