% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_euro_scaled.R
\name{ef_euro_scaled}
\alias{ef_euro_scaled}
\title{Scaled emission factors from European (EMEP/EEA) methodology}
\usage{
ef_euro_scaled(
  dfcol,
  speed,
  veh_type,
  fuel,
  euro,
  SDC = 34.12,
  tech = "SCR",
  pollutant,
  slope = 0,
  load = 0.5,
  show.equation = TRUE
)
}
\arguments{
\item{dfcol}{units; Local emission factors, in units [g/km]}

\item{speed}{units; Speed, in units [km/h]}

\item{veh_type}{character; Bus type, classified in "Urban Buses Midi <=15 t",
"Urban Buses Standard 15 - 18 t", "Urban Buses Articulated >18 t", "Coaches Standard <=18 t"       
"Coaches Articulated >18 t", "Urban CNG Buses", "Urban Biodiesel Buses".}

\item{fuel}{character; Fuel type, classified in "Diesel", "CNG"  and "Biodiesel".}

\item{euro}{character; Euro period of vehicle, classified in "Conventional", "Euro I", "Euro II",
"Euro III", "Euro IV", "Euro V", "Euro VI" and "EEV".}

\item{SDC}{numeric; Speed Driving Cycle (default = 34.12).}

\item{tech}{character; Technology, classified in "SCR", "EGR", "DPF+SCR" or NA (Not available).
For EURO IV, V, default is "SCR". For Euro III, default is NA.}

\item{pollutant}{character; Pollutant, classified in "CO", "NOx", "VOC", "PM", "FC", "CH4", "NH3"
and "N2O".}

\item{slope}{numeric; Slope gradient, classified in -0.06, -0.04, -0.02, 0.00, 0.02, 0.04 and
0.06. Negative gradients means downhills and positive uphills. Default is 0.0.}

\item{load}{numeric; Load ratio, classified in 0.0, 0.5 and 1.0. Default is 0.5.}

\item{show.equation}{logical; show.equation from EMEP/EEEA used? Default parameter is TRUE.}
}
\value{
emission factors in units 'g/km'.
}
\description{
Estimates scaled emission factor for buses based on EMEP/EEA
}
