% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis.R
\name{emis}
\alias{emis}
\title{Estimate hot exhaust emissions}
\usage{
emis(fleet_composition, dist, ef, aggregate = TRUE, prefix)
}
\arguments{
\item{fleet_composition}{vector; Fleet composition.}

\item{dist}{units (in km); Length of each link in km.}

\item{ef}{units (in g/km);    List of functions of emission factors.}

\item{aggregate}{logical; if TRUE (default) emissions are aggregated by pollutant.}

\item{prefix}{character; Add prefix into emissions names. Missing parameter (default)
means empty prefix.}
}
\value{
units (in g); emissions per link.
}
\description{
Calculate hot exhaust emissions.
}
