% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis.R
\name{emis}
\alias{emis}
\title{Estimate hot exhaust emissions}
\usage{
emis(fleet_composition, dist, ef, aggregate = TRUE)
}
\arguments{
\item{fleet_composition}{Fleet composition; Vector.}

\item{dist}{Units (in km); Length of each link in km}

\item{ef}{Units (in g/km);    List of functions of emission factors}

\item{aggregate}{Logical; if TRUE (default) emissions are aggregated by pollutant}
}
\value{
Units (in g); emissions per link
}
\description{
Calculate hot exhaust emissions
}
