% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis.R
\name{emis}
\alias{emis}
\title{Estimate hot exhaust emissions}
\usage{
emis(veh, dist, ef)
}
\arguments{
\item{veh}{Vehicles data-frame or list of "Vehicles" data-frame.
Each data-frame as number of columns matching the age distribution of that ype of vehicle. The number of rows is equal to the number of streets link. If this is a list, the length of the list is the vehicles for each hour.}

\item{dist}{Units (in km); Length of each link in km}

\item{ef}{Units (in g/km);    List of functions of emission factors}
}
\value{
Units (in g); emissions per link
}
\description{
Calculate hot exhaust emissions
}
