% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis.R
\name{emis}
\alias{emis}
\title{Estimate hot exhaust emissions}
\usage{
emis(
  gtfs_data,
  fleet_composition,
  ef,
  aggregate = TRUE,
  prefix = NULL,
  as_list = TRUE
)
}
\arguments{
\item{gtfs_data}{A path to a GTFS.zip file, or a GTFS data represented as a 
list of data.tables generated by the `gtfs2gps` package.}

\item{fleet_composition}{vector; Fleet composition, which is a distribution 
of fleet based on frequency. If there is only one.}

\item{ef}{list or data.table;    Emission factors.}

\item{aggregate}{logical; if TRUE (default) emissions are aggregated by pollutant.}

\item{prefix}{character; Add prefix into emissions names. Missing parameter (default)
means empty prefix.}

\item{as_list}{logical; if TRUE (default) emissions are returned inside 'ef' list.}
}
\value{
units ('g'); emissions per link.
}
\description{
Calculate hot exhaust emissions.
}
\examples{

library(gtfs2emis)
library(data.table)
library(magrittr)

gtfs <- read_gtfs(system.file("extdata/gtfs_cur.zip", package = "gtfs2emis")) \%>\%
              filter_by_shape_id("1708") \%>\%
                 filter_single_trip()


# set.seed(1335)
# dist = units::set_units(rnorm(100,0.250,0.03),"km")
# ef <- ef_europe(speed = units::set_units(rnorm(100,50,5),"km/h"),
#                 veh_type = c("Urban Buses Standard 15 - 18 t","Urban Buses Articulated >18 t"),
#                 euro = c("IV","V"),
#                 pollutant = c("CO2","NOx"),
#                 fuel = "Diesel" ,
#                 tech =  c("SCR","EGR"),
#                 slope = 0.0,
#                 load = 0.5,
#                 fcorr = 1,
#                 as_list = TRUE)
# 
# emi <- emis(fleet_composition =  c(0.7,0.3),
#             dist = dist,
#             ef = ef,
#             aggregate = FALSE,
#             as_list = TRUE)  
            
}
