% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_grid.R
\name{emis_grid}
\alias{emis_grid}
\title{Allocate emissions into a spatial grid}
\usage{
emis_grid(data, emi, grid, time_class, time_column)
}
\arguments{
\item{data}{Data.frame; Data.frame containing the emissions and time stamp column data.}

\item{emi}{Character; Column names of emissions in 'data'.}

\item{grid}{Sf polygon; Grid cell data to allocate emissions.}

\item{time_class}{Character; type of time aggregation in 'data', which can be
an aggregation by 'all periods' (Default), 'hour' or 'hour-minute'.}

\item{time_column}{Vector; Column name with time stamp information (from 'data').
Argument required when time_class = 'hour' or 'hour-minute' is selected.}
}
\value{
An \verb{"sf" "data.frame"} object with emissions estimates per grid cell.
}
\description{
Aggregate emissions proportionally in an sf polygon grid, by performing an
intersection operation between emissions data in \verb{sf linestring} format and
the input grid cells. User can also aggregate the emissions in the grid
by time of the day.
}
\examples{
if (interactive()) {
# GTFS2gps filter-----
library(data.table)
library(magrittr)
fort <- gtfs2gps::read_gtfs(system.file("extdata/fortaleza.zip"
                                       , package = "gtfs2gps"))  \%>\%
 gtfs2gps::filter_single_trip() \%>\% 
 gtfs2gps::filter_by_shape_id(c("shape804-I", "shape806-I"))

fort_gps <- gtfs2gps::gtfs2gps(fort, parallel = TRUE)

fort_gpslines <- gtfs2gps::gps_as_sflinestring(fort_gps)

fort_gpslines$dist <- units::set_units(fort_gpslines$dist, "km")

# fleet -----------
total_fleet <- data.table::data.table(year = c(2005,2010,2011,2012
                                              ,2014,2015,2017,2018,2019),
                                     bus = c(1,61,50,1,45,18,62,27,31),
                                     veh_type_euro = "Ubus Std 15 - 18 t",
                                     euro_stage = c("II", "IV", "IV", "V"
                                                    , "V", "V", "V", "V","V"))
total_fleet$fleet_composition <- total_fleet$bus/sum(total_fleet$bus)

# Emission factor
set.seed(1234)
EF_europe <- ef_europe_emep(pollutant = c("CO", "PM10"),
                      speed = fort_gpslines$speed,
                      veh_type = total_fleet$veh_type_euro,
                      tech = "SCR",
                      euro = total_fleet$euro_stage,
                      fcorr = rnorm(9, 0.5, 0.1))
# EUROPE
emi_europe <- emis(fleet_composition = total_fleet$fleet_composition,
                  dist = fort_gpslines$dist,
                  ef = EF_europe,
                  prefix = "EU")

for_emis <- cbind(fort_gpslines, emi_europe$emi)

# Grid
grid_gps <- sf::st_make_grid(for_emis, cellsize = 0.25 / 102.47, square = FALSE)
for_sf <- sf::st_as_sf(for_emis)

pol_gps <- emis_grid(data = for_sf,
                    emi = c("EU_CO_total"),
                    grid = grid_gps,
                    time_class = 'all periods')

pol_gps_hour <- emis_grid(data = for_sf,
                         emi = c("EU_CO_total"),
                         grid = grid_gps,
                         time_class = 'hour',
                         time_column = 'timestamp')

pol_gps_hour_minute <- emis_grid(data = for_sf,
                                emi = c("EU_CO_total"),
                                grid = grid_gps,
                                time_class = 'hour-minute',
                                time_column = 'timestamp')
}
}
