% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_grid.R
\name{emis_grid}
\alias{emis_grid}
\title{Street emissions into grid}
\usage{
emis_grid(data, emi, grid, time_class, time_column)
}
\arguments{
\item{data}{data.table; Data.table with emissions and departure_time data.}

\item{emi}{character; Columns names of emissions information.}

\item{grid}{sf polygon; Grid cell data to allocate emissions.}

\item{time_class}{character; Emissions can be aggregated by 'all periods' (Default),
'hour' or 'hour-minute'.}

\item{time_column}{vector; Column name of 'departure_time' information (from data input). 
Only used when 'hour' or 'hour-minute' is selected.}
}
\value{
An `"sf" "data.frame"` object with emissions estimates per grid cell.
}
\description{
Aggregate emissions proportionally into grid cells, by performing an 
intersection operation between emissions data in `sf linestring` format and 
grid cells. It also considers aggregation per grid cells and time.
}
\examples{
# GTFS2gps filter-----
library(data.table)
library(magrittr)
fort <- gtfs2gps::read_gtfs(system.file("extdata/fortaleza.zip"
                                       , package = "gtfs2gps"))  \%>\%
 gtfs2gps::filter_single_trip() \%>\% 
 gtfs2gps::filter_by_shape_id(c("shape804-I", "shape806-I"))

fort_gps <- gtfs2gps::gtfs2gps(fort, parallel = TRUE)

fort_gpslines <- gtfs2gps::gps_as_sflinestring(fort_gps)

fort_gpslines$dist <- units::set_units(fort_gpslines$dist, "km")

# fleet -----------
total_fleet <- data.table::data.table(year = c(2005,2010,2011,2012
                                              ,2014,2015,2017,2018,2019),
                                     bus = c(1,61,50,1,45,18,62,27,31),
                                     veh_type_euro = "Ubus Std 15 - 18 t",
                                     euro_stage = c("II", "IV", "IV", "V"
                                                    , "V", "V", "V", "V","V"))
total_fleet$fleet_composition <- total_fleet$bus/sum(total_fleet$bus)

# Emission factor
set.seed(1234)
EF_europe <- ef_europe(pollutant = c("CO", "PM10"),
                      speed = fort_gpslines$speed,
                      veh_type = total_fleet$veh_type_euro,
                      tech = "SCR",
                      euro = total_fleet$euro_stage,
                      fcorr = rnorm(9, 0.5, 0.1))
# EUROPE
emi_europe <- emis(fleet_composition = total_fleet$fleet_composition,
                  dist = fort_gpslines$dist,
                  ef = EF_europe,
                  prefix = "EU")

for_emis <- cbind(fort_gpslines, emi_europe$emi)

# Grid
grid_gps <- vein::make_grid(spobj = for_emis, width =  0.25 / 102.47) # 500 meters
for_sf <- sf::st_as_sf(for_emis)

pol_gps <- emis_grid(data = for_sf,
                    emi = c("EU_CO_total"),
                    grid = grid_gps,
                    time_class = 'all periods')

pol_gps_hour <- emis_grid(data = for_sf,
                         emi = c("EU_CO_total"),
                         grid = grid_gps,
                         time_class = 'hour',
                         time_column = 'timestamp')

pol_gps_hour_minute <- emis_grid(data = for_sf,
                                emi = c("EU_CO_total"),
                                grid = grid_gps,
                                time_class = 'hour-minute',
                                time_column = 'timestamp')

}
