% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_post.R
\name{emis_post}
\alias{emis_post}
\title{Emissions post-processsing}
\usage{
emis_post(emi, veh, time, time_dt)
}
\arguments{
\item{emi}{List; Emissions (in units 'g') generated by function emis}

\item{veh}{Character; Vehicle type displayed in 'names(emi)', 'all' 
(full aggregation by 'vehicle_type'), or 'none' Missing input the emissions are not aggregated.}

\item{time}{Character; Emission can be aggregated into 'hour' and'hour-minute'. 
Missing input (Default) no aggregation without considering time consideration.}

\item{time_dt}{Vector; Vector containing 'departure_time'. 
Only used when 'hour' or 'hour-minute' is selected.}
}
\value{
Units (in g); emissions
}
\description{
Aggregate emissions generated by emis function by time, 
vehicle type and segment link (spatial)
}
