% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_post.R
\name{emis_post}
\alias{emis_post}
\title{Emissions post-processing}
\usage{
emis_post(data, emi, time_class, time_column = "departure_time")
}
\arguments{
\item{data}{data.table; Data.table with emissions and departure_time data.}

\item{emi}{character; Columns names of emissions information.}

\item{time_class}{character; Emissions can be aggregated by 'hour' or 'hour-minute'.}

\item{time_column}{vector; Vector containing 'departure_time'. 
Only used when 'hour' or 'hour-minute' is selected.}
}
\value{
units ('g'); emissions.
}
\description{
Aggregate emissions generated by emis function by time, 
vehicle type and segment link (spatial).
}
\examples{
require(magrittr)
data <- gtfs2gps::read_gtfs(system.file("extdata/saopaulo.zip", package = "gtfs2gps")) \%>\%
  gtfs2gps::filter_by_shape_id("51982") \%>\%
  gtfs2gps::gtfs2gps() \%>\%
  gtfs2gps::gps_as_sflinestring() \%>\%
  dplyr::select(trip_id, speed, dist, departure_time)

data$dist <- units::set_units(data$dist, "km")

total_fleet <- data.table::data.table(year = c(2005, 2010:2012, 2014, 2015, 2017:2019),
                                      bus = c(1, 61, 50, 1, 45, 18, 62, 27, 31),
                                      veh_type_euro = "Urban Buses Standard 15 - 18 t",
                                      euro_stage = c("II", "IV", "IV", rep("V", 6)))
total_fleet$fleet_composition = total_fleet$bus/sum(total_fleet$bus)

EF_usa <-  gtfs2emis::ef_usa(pollutant = c("CO", "PM10"),
                 calendar_year = "2019",
                 model_year = total_fleet$year,
                 speed = data$speed,
                 fuel = "Diesel")
emi_usa <- gtfs2emis::emis(fleet_composition = total_fleet$fleet_composition,
                           dist = data$dist, ef = EF_usa, prefix = "USA")
data <- cbind(data, emi_usa)

emi_post_usa <- gtfs2emis::emis_post(data = data, time_class = "hour",
                                     emi = c("USA_CO_total", "USA_PM10_total"),
                                      time_column = "departure_time")
}
