% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_post.R
\name{emis_post}
\alias{emis_post}
\title{Emissions post-processsing}
\usage{
emis_post(emi, veh, time, time_dt)
}
\arguments{
\item{emi}{List; Emissions (in units 'g') generated by function emis}

\item{veh}{Character; Vehicle type displayed in 'names(emi)', 'all' 
(full aggregation by 'vehicle_type'), or 'none' Missing input the emissions are not aggregated.}

\item{time}{Character; Emission can be aggregated into 'hour' and'hour-minute'. 
Missing input (Default) no aggregation without considering time consideration.}

\item{time_dt}{Vector; Vector containing 'departure_time'. 
Only used when 'hour' or 'hour-minute' is selected.}
}
\value{
Units (in g); emissions
}
\description{
Aggregate emissions generated by emis function by time, 
vehicle type and segment link (spatial)
}
\examples{
set.seed(12039)
temp_emi <- emis(veh = data.table::data.table("veh1" = 0.75,"veh2" = 0.25), # sum(veh) = 1
                 dist = units::set_units(rnorm(10,mean = 8),'km'),
                 ef = list("veh1" = units::set_units(rnorm(10,0.35,0.01),'g/km'),
                           "veh2" = units::set_units(rnorm(10,0.27,0.01),'g/km')))
temp_time <- c("10:40","10:40","10:42","10:44","10:45",
               "10:46","10:47","10:48","10:49","10:50")

emis_post(emi = temp_emi,veh = 'all',time = 'hour',time_dt = temp_time)
emis_post(emi = temp_emi,veh = 'all',time = 'hour-minute',time_dt = temp_time)
emis_post(emi = temp_emi,veh = 'veh1',time = 'hour-minute',time_dt = temp_time)
emis_post(emi = temp_emi,time = 'hour',time_dt = temp_time)
emis_post(emi = temp_emi,veh = 'veh1')
emis_post(emi = temp_emi)
}
