% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_summary.R
\name{emis_summary}
\alias{emis_summary}
\title{Emissions summary by aggregating results}
\usage{
emis_summary(
  emi,
  emi_vars,
  by,
  time_column = NULL,
  pol_vars = NULL,
  veh_vars = NULL
)
}
\arguments{
\item{emi}{data.table; Data.table with emissions and departure_time data.}

\item{emi_vars}{character; Columns names of emissions information.}

\item{by}{character; Emissions can be aggregated by 'time', 'age', 'veh_type' and 'pollutant'}

\item{time_column}{vector; Vector containing time_column.}

\item{pol_vars}{vector; Vector containing variables of pollutant. 
Only used when 'hour' or 'hour-minute' is selected.}

\item{veh_vars}{vector; Vector containing variables of vehicle type.}
}
\value{
units ('g'); emissions.
}
\description{
Aggregate emissions generated by the \code{\link{emis}} function by time, 
vehicle type and segment link (spatial).
}
\examples{
set.seed(1335)
dist = units::set_units(rnorm(100,0.250,0.03),"km")
ef <- ef_europe(speed = units::set_units(rnorm(100,50,5),"km/h"),
                veh_type = c("Urban Buses Standard 15 - 18 t","Urban Buses Articulated >18 t"),
                euro = c("IV","V"),
                pollutant = c("CO2","NOx"),
                fuel = "Diesel" ,
                tech =  c("SCR","EGR"),
                slope = 0.0,
                load = 0.5,
                fcorr = 1,
                as_list = TRUE)

emi <- emis(fleet_composition =  c(0.7,0.3),
            dist = dist,
            ef = ef,
            aggregate = FALSE,
            as_list = TRUE)  

time_column <- data.table::as.ITime(seq(1,14400,length.out = 100))
my_emis_summary <- emis_summary(emi = emi,
                     emi_var = "emi", 
                     by = "time",  # veh_type, time, pollutant 
                     time_column = time_column,
                     veh_var = "veh_type", # veh_type
                     pol_var = "pollutant") # pollutant
}
