% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission_model.R
\name{emission_model}
\alias{emission_model}
\title{Emissions model}
\usage{
emission_model(
  gps_line_path,
  ef_data_base,
  fleet_data,
  emis_path,
  pollutant,
  parallel,
  calendar_year,
  is_ef_scaled
)
}
\arguments{
\item{gps_line_path}{Character; Filepath with the GPS files as sf_linestring.}

\item{ef_data_base}{Character; Emission factor database, such as "moves_usa",
"emfac_usa", "emep_europe","cetesb_brazil".}

\item{fleet_data}{Data.table. Database with fleet_data. The requires columns 
depends on the ef_data_base source. See @details for input examples.}

\item{emis_path}{Character; Filepath to export emissions.}

\item{pollutant}{Vector; Vector of characters of pollutants. character; The available 
pollutants depends on the ef_data_base. See `ef_usa_moves`,`ef_usa_emfac`,
`ef_brazil_cetesb` or `ef_emep_europe` for details.}

\item{parallel}{Logical; Decides whether the function should run in parallel. 
Defaults is FALSE.When TRUE, it will use all cores available minus one 
using future::plan() with strategy "multisession" internally.}

\item{load}{Numeric; Load ratio, classified in 0.0, 0.5 and 1.0. Default is 0.5.
This is required only `ef_emep_europe` is selected.}

\item{as_list}{Logical; Returns emission factors as a list, instead of 
data.table format. Default is TRUE.}

\item{workers}{Numeric; Number of workers available for parallel processing.}

\item{tech}{character; After treatment technology, classified in "SCR"
(Selective Catalytic Reduction),}
}
\description{
Estimate hot-exhaust emissions from the input transport model
}
\details{
In the fleet database..
- Column `calendar_year`: Character; Base year of the @ef_data_base input. 
Required only when  `ef_usa_moves` or `ef_usa_emfac` are selected.
- Column `tech`  Character; After treatment technology, classified in "SCR" 
(Selective Catalytic Reduction), "EGR" (Exhaust Gas Recirculation), and 
"DPF+SCR" (Diesel Particulate Filter + SCR, for Euro VI). Default is "SCR" 
for "IV" and "V". This is required only when `ef_emep_europe` or 
`ef_euro_scaled` are selected.
- Column `euro`  character; Euro period of vehicle, classified in 
"Conventional", "I", "II", "III", "IV", "V", "VI", and "EEV". This is required 
only when `ef_emep_europe` or `ef_euro_scaled` are selected.
- Column `fuel` character; Fuel type, classified in "D" (Diesel)
,"DHD" (Diesel Hybrid ~ Diesel), "DHE" (Diesel Hybrid ~ Electricity)
, "CNG" (Compressed Natural Gas), "BD" (Biodiesel).
}
