% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps2line.R
\name{gps2line}
\alias{gps2line}
\title{Convert GPS to Lines-data type}
\usage{
gps2line(input_file, output_filepath = NA)
}
\arguments{
\item{input_file}{Data.frame/Data.table; location of exported GPS files or GPS files}

\item{output_filepath}{Character; location for export Line-data type. If missing it returns the values.}
}
\description{
Transform intervals of GPS data of constant speed into
a linestring segment, in order to reduce data size of observations.
}
