% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps2line.R
\name{gps_as_sflinestring}
\alias{gps_as_sflinestring}
\title{Converts a GPS-like data.table to a MultiLineString Simple Feature (sf) object}
\usage{
gps_as_sflinestring(input_file, crs = 4326)
}
\arguments{
\item{input_file}{A data.table with timestamp data.}

\item{crs}{A Coordinate Reference System. The default value is 4326 (latlong WGS84).}
}
\value{
A simple feature (sf) object with MultiLineString data.
}
\description{
Every interval of GPS data points of constant speed for each trip_id is
 converted into a linestring segment.
}
\examples{
library(gtfs2gps)
library(gtfs2emis)

cur <- gtfs2gps::gtfs2gps(system.file("extdata/inst/gtfs_cur.zip", package = "gtfs2emis"))
subset <- filter_by_shape_id(poa, "T2-1") \%>\%
  filter_single_trip()

poa_gps <- gtfs2gps(subset)
poa_gps_sf <- gps_as_sf(poa_gps)

}
