% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_emfac_db}
\alias{usa_emfac_db}
\title{EMFAC
Running exhaust emissions factors from EMFAC2017 model}
\format{
A data.table with 106920 rows and 7 variables:
\describe{
\item{pol}{Character; Pollutants: Carbon monoxide (CO); Nitrogen oxides (NOx); 
  Hydrocarbons as TOG (total organic gases), ROG (reactive organic gases), 
  THC (total hydrocarbon), or CH4 (methane); 
  Particulate matter as particulate matters 10 microns or less in diameter (PM10), 
  and particulate matters 2.5 microns or less in diameter (PM2.5);
  Sulfur oxides (SOx); Carbon Dioxide (CO2); 
  Nitrous Oxide (N2O) and Methane (CH4).}
  \item{calendar_year}{Numeric; Calendar Year between 2010 - 2020}
  \item{fuel}{character; Type of fuel: 'D' (Diesel),'G' (Gasoline),'CNG' (Compressed Natural Gas).}
  \item{model_year}{Model year.}
  \item{speed}{Units; Speed in 'km/h'; Emission factor are returned in speed intervals 
 such as "5-10", "10-15", "15-20", "20-25", "25-30", "30-35", "35-40", "40-45", "45-50"
  "50-55", "55-60", "60-65", "65-70", "70-75", "75-80", "80-85", "85-90", ">90" mph (miles/h)}
}
}
\source{
\url{https://www.eea.europa.eu/publications/emep-eea-guidebook-2019/part-b-sectoral-guidance-chapters/1-energy/1-a-combustion/road-transport-appendix-4-emission/view}
}
\usage{
usa_emfac_db
}
\description{
EMFAC
Running exhaust emissions factors from EMFAC2017 model
}
\keyword{datasets}
